@@ -4,6 +4,12 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "type": "project",
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "https://github.com/schooch/twitter"
+        }
+    ],
     "require": {
         "php": ">=7.0.0",
         "aws/aws-sdk-php-laravel": "~3.0",
@@ -15,39 +21,38 @@
         "doctrine/dbal": "^2.6",
         "dumpk/elastcoder": "^0.9.1",
         "facebook/graph-sdk": "^5.6",
-        "fideloper/proxy": "~3.3",
+        "fideloper/proxy": "~4.0",
         "google/apiclient": "^2.2",
         "google/cloud-storage": "^1.0",
         "intervention/image": "^2.4",
         "janyksteenbeek/youtubepartner": "^1.1",
         "kouz/laravel-airbrake": "^0.3.0",
-        "laravel/framework": "5.5.*",
+        "laravel/framework": "5.6.*",
         "laravel/tinker": "~1.0",
         "laravelcollective/html": "^5.5",
         "league/csv": "^9.0",
         "league/flysystem-aws-s3-v3": "^1.0",
         "michaeljwright/aws-rekognition": "^0.2",
         "nao-pon/flysystem-google-drive": "~1.1",
-        "pbmedia/laravel-ffmpeg": "^1.3.0",
+        "pbmedia/laravel-ffmpeg": "^2.0",
         "predis/predis": "^1.1",
         "ralouphie/mimey": "^1.0",
         "ramsey/uuid": "^3.7",
         "schooch/youtube": "^5.0.5",
         "talvbansal/media-manager": "^2.5",
-        "thujohn/twitter": "^2.2",
+        "thujohn/twitter": "dev-test",
         "true/punycode": "~2.0",
         "weidner/goutte": "^1.0"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.1",
         "barryvdh/laravel-ide-helper": "^2.4",
+        "codeception/codeception": "^2.4",
         "filp/whoops": "~2.0",
         "fzaninotto/faker": "~1.4",
-        "google/cloud-tools": "^0.6",
-        "laravel/dusk": "^2.0",
+        "laravel/dusk": "~4.0",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~6.0",
-        "symfony/browser-kit": "^3.0"
+        "phpunit/phpunit": "~7.0"
     },
     "autoload": {
         "classmap": [
