@@ -2,165 +2,163 @@
 
 return [
 
-    'name' => env('APP_NAME', 'Sniffr'),
-    'logo' => env('APP_LOGO', ''),
-    'env' => env('APP_ENV', 'prod'),
-    'debug' => env('APP_DEBUG', false),
-    'url' => env('APP_URL', 'http://localhost'),
-    'timezone' => 'Europe/London',
-    'locale' => 'en',
-    'fallback_locale' => 'en',
-    'key' => env('APP_KEY'),
-    'cipher' => 'AES-256-CBC',
+	'name' => env('APP_NAME', 'Sniffr'),
+	'logo' => env('APP_LOGO', ''),
+	'env' => env('APP_ENV', 'prod'),
+	'debug' => env('APP_DEBUG', false),
+	'url' => env('APP_URL', 'http://localhost'),
+	'timezone' => 'Europe/London',
+	'locale' => 'en',
+	'fallback_locale' => 'en',
+	'key' => env('APP_KEY'),
+	'cipher' => 'AES-256-CBC',
 
-    /*
-    |--------------------------------------------------------------------------
-    | Logging Configuration
-    |--------------------------------------------------------------------------
-    |
-    | Here you may configure the log settings for your application. Out of
-    | the box, Laravel uses the Monolog PHP logging library. This gives
-    | you a variety of powerful log handlers / formatters to utilize.
-    |
-    | Available Settings: "single", "daily", "syslog", "errorlog"
-    |
-    */
+	/*
+	|--------------------------------------------------------------------------
+	| Logging Configuration
+	|--------------------------------------------------------------------------
+	|
+	| Here you may configure the log settings for your application. Out of
+	| the box, Laravel uses the Monolog PHP logging library. This gives
+	| you a variety of powerful log handlers / formatters to utilize.
+	|
+	| Available Settings: "single", "daily", "syslog", "errorlog"
+	|
+	*/
 
-    'log' => env('APP_LOG', 'single'),
+	'log' => env('APP_LOG', 'single'),
 
-    'log_level' => env('APP_LOG_LEVEL', 'debug'),
+	'log_level' => env('APP_LOG_LEVEL', 'debug'),
 
-    /*
-    |--------------------------------------------------------------------------
-    | Autoloaded Service Providers
-    |--------------------------------------------------------------------------
-    |
-    | The service providers listed here will be automatically loaded on the
-    | request to your application. Feel free to add your own services to
-    | this array to grant expanded functionality to your applications.
-    |
-    */
+	/*
+	|--------------------------------------------------------------------------
+	| Autoloaded Service Providers
+	|--------------------------------------------------------------------------
+	|
+	| The service providers listed here will be automatically loaded on the
+	| request to your application. Feel free to add your own services to
+	| this array to grant expanded functionality to your applications.
+	|
+	*/
 
-    'providers' => [
+	'providers' => [
 
-        /*
-         * Laravel Framework Service Providers...
-         */
-        Illuminate\Auth\AuthServiceProvider::class,
-        Illuminate\Broadcasting\BroadcastServiceProvider::class,
-        Illuminate\Bus\BusServiceProvider::class,
-        Illuminate\Cache\CacheServiceProvider::class,
-        Illuminate\Foundation\Providers\ConsoleSupportServiceProvider::class,
-        Illuminate\Cookie\CookieServiceProvider::class,
-        Illuminate\Database\DatabaseServiceProvider::class,
-        Illuminate\Encryption\EncryptionServiceProvider::class,
-        Illuminate\Filesystem\FilesystemServiceProvider::class,
-        Illuminate\Foundation\Providers\FoundationServiceProvider::class,
-        Illuminate\Hashing\HashServiceProvider::class,
-        Illuminate\Mail\MailServiceProvider::class,
-        Illuminate\Notifications\NotificationServiceProvider::class,
-        Illuminate\Pagination\PaginationServiceProvider::class,
-        Illuminate\Pipeline\PipelineServiceProvider::class,
-        Illuminate\Queue\QueueServiceProvider::class,
-        Illuminate\Redis\RedisServiceProvider::class,
-        Illuminate\Auth\Passwords\PasswordResetServiceProvider::class,
-        Illuminate\Session\SessionServiceProvider::class,
-        Illuminate\Translation\TranslationServiceProvider::class,
-        Illuminate\Validation\ValidationServiceProvider::class,
-        Illuminate\View\ViewServiceProvider::class,
+		/*
+		 * Laravel Framework Service Providers...
+		 */
+		Illuminate\Auth\AuthServiceProvider::class,
+		Illuminate\Broadcasting\BroadcastServiceProvider::class,
+		Illuminate\Bus\BusServiceProvider::class,
+		Illuminate\Cache\CacheServiceProvider::class,
+		Illuminate\Foundation\Providers\ConsoleSupportServiceProvider::class,
+		Illuminate\Cookie\CookieServiceProvider::class,
+		Illuminate\Database\DatabaseServiceProvider::class,
+		Illuminate\Encryption\EncryptionServiceProvider::class,
+		Illuminate\Filesystem\FilesystemServiceProvider::class,
+		Illuminate\Foundation\Providers\FoundationServiceProvider::class,
+		Illuminate\Hashing\HashServiceProvider::class,
+		Illuminate\Mail\MailServiceProvider::class,
+		Illuminate\Notifications\NotificationServiceProvider::class,
+		Illuminate\Pagination\PaginationServiceProvider::class,
+		Illuminate\Pipeline\PipelineServiceProvider::class,
+		Illuminate\Queue\QueueServiceProvider::class,
+		Illuminate\Redis\RedisServiceProvider::class,
+		Illuminate\Auth\Passwords\PasswordResetServiceProvider::class,
+		Illuminate\Session\SessionServiceProvider::class,
+		Illuminate\Translation\TranslationServiceProvider::class,
+		Illuminate\Validation\ValidationServiceProvider::class,
+		Illuminate\View\ViewServiceProvider::class,
 
-        /*
-         * Package Service Providers...
-         */
-        Intervention\Image\ImageServiceProvider::class,
-        Aws\Laravel\AwsServiceProvider::class,
-        Collective\Html\HtmlServiceProvider::class,
-        Schooch\Youtube\YoutubeServiceProvider::class,
-        BaoPham\DynamoDb\DynamoDbServiceProvider::class,
-        Pbmedia\LaravelFFMpeg\FFMpegServiceProvider::class,
-        Intervention\Image\ImageServiceProvider::class,
-        Thujohn\Twitter\TwitterServiceProvider::class,
-        Kouz\Providers\AirbrakeServiceProvider::class,
-        Barryvdh\DomPDF\ServiceProvider::class,
-        MichaelJWright\Rekognition\RekognitionServiceProvider::class,
-        \TalvBansal\MediaManager\Providers\MediaManagerServiceProvider::class,
-        ComoCode\LaravelAb\LaravelAbServiceProvider::class,
-        Weidner\Goutte\GoutteServiceProvider::class,
+		/*
+		 * Package Service Providers...
+		 */
+		Intervention\Image\ImageServiceProvider::class,
+		Aws\Laravel\AwsServiceProvider::class,
+		Collective\Html\HtmlServiceProvider::class,
+		Schooch\Youtube\YoutubeServiceProvider::class,
+		BaoPham\DynamoDb\DynamoDbServiceProvider::class,
+		Pbmedia\LaravelFFMpeg\FFMpegServiceProvider::class,
+		Thujohn\Twitter\TwitterServiceProvider::class,
+		Kouz\Providers\AirbrakeServiceProvider::class,
+		Barryvdh\DomPDF\ServiceProvider::class,
+		MichaelJWright\Rekognition\RekognitionServiceProvider::class,
+		\TalvBansal\MediaManager\Providers\MediaManagerServiceProvider::class,
+		ComoCode\LaravelAb\LaravelAbServiceProvider::class,
+		Weidner\Goutte\GoutteServiceProvider::class,
 		CodeWizz\RedditAPI\RedditAPIServiceProvider::class,
 
-        /*
-         * Application Service Providers...
-         */
-        App\Providers\AppServiceProvider::class,
-        App\Providers\AuthServiceProvider::class,
-        // App\Providers\BroadcastServiceProvider::class,
-        App\Providers\EventServiceProvider::class,
-        App\Providers\RouteServiceProvider::class,
-        App\Providers\GoogleDriveServiceProvider::class,
-        Chumper\Zipper\ZipperServiceProvider::class,
+		/*
+		 * Application Service Providers...
+		 */
+		App\Providers\AppServiceProvider::class,
+		App\Providers\AuthServiceProvider::class,
+		App\Providers\EventServiceProvider::class,
+		App\Providers\RouteServiceProvider::class,
+		App\Providers\GoogleDriveServiceProvider::class,
+		Chumper\Zipper\ZipperServiceProvider::class,
 		Barryvdh\Debugbar\ServiceProvider::class
-    ],
+	],
 
-    /*
-    |--------------------------------------------------------------------------
-    | Class Aliases
-    |--------------------------------------------------------------------------
-    |
-    | This array of class aliases will be registered when this application
-    | is started. However, feel free to register as many as you wish as
-    | the aliases are "lazy" loaded so they don't hinder performance.
-    |
-    */
+	/*
+	|--------------------------------------------------------------------------
+	| Class Aliases
+	|--------------------------------------------------------------------------
+	|
+	| This array of class aliases will be registered when this application
+	| is started. However, feel free to register as many as you wish as
+	| the aliases are "lazy" loaded so they don't hinder performance.
+	|
+	*/
 
-    'aliases' => [
+	'aliases' => [
 
-        'App' => Illuminate\Support\Facades\App::class,
-        'Artisan' => Illuminate\Support\Facades\Artisan::class,
-        'Auth' => Illuminate\Support\Facades\Auth::class,
-        'Blade' => Illuminate\Support\Facades\Blade::class,
-        'Broadcast' => Illuminate\Support\Facades\Broadcast::class,
-        'Bus' => Illuminate\Support\Facades\Bus::class,
-        'Cache' => Illuminate\Support\Facades\Cache::class,
-        'Config' => Illuminate\Support\Facades\Config::class,
-        'Cookie' => Illuminate\Support\Facades\Cookie::class,
-        'Crypt' => Illuminate\Support\Facades\Crypt::class,
-        'DB' => Illuminate\Support\Facades\DB::class,
-        'Eloquent' => Illuminate\Database\Eloquent\Model::class,
-        'Event' => Illuminate\Support\Facades\Event::class,
-        'File' => Illuminate\Support\Facades\File::class,
-        'Gate' => Illuminate\Support\Facades\Gate::class,
-        'Hash' => Illuminate\Support\Facades\Hash::class,
-        'Lang' => Illuminate\Support\Facades\Lang::class,
-        'Log' => Illuminate\Support\Facades\Log::class,
-        'Mail' => Illuminate\Support\Facades\Mail::class,
-        'Notification' => Illuminate\Support\Facades\Notification::class,
-        'Password' => Illuminate\Support\Facades\Password::class,
-        'Queue' => Illuminate\Support\Facades\Queue::class,
-        'Redirect' => Illuminate\Support\Facades\Redirect::class,
-        'Redis' => Illuminate\Support\Facades\Redis::class,
-        'Request' => Illuminate\Support\Facades\Request::class,
-        'Response' => Illuminate\Support\Facades\Response::class,
-        'Route' => Illuminate\Support\Facades\Route::class,
-        'Schema' => Illuminate\Support\Facades\Schema::class,
-        'Session' => Illuminate\Support\Facades\Session::class,
-        'Storage' => Illuminate\Support\Facades\Storage::class,
-        'URL' => Illuminate\Support\Facades\URL::class,
-        'Validator' => Illuminate\Support\Facades\Validator::class,
-        'View' => Illuminate\Support\Facades\View::class,
-        'Image' => Intervention\Image\Facades\Image::class,
-        'Form' => Collective\Html\FormFacade::class,
-        'Html' => Collective\Html\HtmlFacade::class,
-        'AWS' => Aws\Laravel\AwsFacade::class,
-        'Youtube' => Schooch\Youtube\Facades\Youtube::class,
-        'FFMpeg' => Pbmedia\LaravelFFMpeg\FFMpegFacade::class,
-        'Image' => Intervention\Image\Facades\Image::class,
-        'Twitter' => Thujohn\Twitter\Facades\Twitter::class,
-        'PDF' => Barryvdh\DomPDF\Facade::class,
-        'Rekognition' => 'MichaelJWright\Rekognition\RekognitionFacade',
-        'Zipper' => Chumper\Zipper\Zipper::class,
-        'Goutte' => Weidner\Goutte\GoutteFacade::class,
+		'App' => Illuminate\Support\Facades\App::class,
+		'Artisan' => Illuminate\Support\Facades\Artisan::class,
+		'Auth' => Illuminate\Support\Facades\Auth::class,
+		'Blade' => Illuminate\Support\Facades\Blade::class,
+		'Broadcast' => Illuminate\Support\Facades\Broadcast::class,
+		'Bus' => Illuminate\Support\Facades\Bus::class,
+		'Cache' => Illuminate\Support\Facades\Cache::class,
+		'Config' => Illuminate\Support\Facades\Config::class,
+		'Cookie' => Illuminate\Support\Facades\Cookie::class,
+		'Crypt' => Illuminate\Support\Facades\Crypt::class,
+		'DB' => Illuminate\Support\Facades\DB::class,
+		'Eloquent' => Illuminate\Database\Eloquent\Model::class,
+		'Event' => Illuminate\Support\Facades\Event::class,
+		'File' => Illuminate\Support\Facades\File::class,
+		'Gate' => Illuminate\Support\Facades\Gate::class,
+		'Hash' => Illuminate\Support\Facades\Hash::class,
+		'Lang' => Illuminate\Support\Facades\Lang::class,
+		'Log' => Illuminate\Support\Facades\Log::class,
+		'Mail' => Illuminate\Support\Facades\Mail::class,
+		'Notification' => Illuminate\Support\Facades\Notification::class,
+		'Password' => Illuminate\Support\Facades\Password::class,
+		'Queue' => Illuminate\Support\Facades\Queue::class,
+		'Redirect' => Illuminate\Support\Facades\Redirect::class,
+		'Redis' => Illuminate\Support\Facades\Redis::class,
+		'Request' => Illuminate\Support\Facades\Request::class,
+		'Response' => Illuminate\Support\Facades\Response::class,
+		'Route' => Illuminate\Support\Facades\Route::class,
+		'Schema' => Illuminate\Support\Facades\Schema::class,
+		'Session' => Illuminate\Support\Facades\Session::class,
+		'Storage' => Illuminate\Support\Facades\Storage::class,
+		'URL' => Illuminate\Support\Facades\URL::class,
+		'Validator' => Illuminate\Support\Facades\Validator::class,
+		'View' => Illuminate\Support\Facades\View::class,
+		'Image' => Intervention\Image\Facades\Image::class,
+		'Form' => Collective\Html\FormFacade::class,
+		'Html' => Collective\Html\HtmlFacade::class,
+		'AWS' => Aws\Laravel\AwsFacade::class,
+		'Youtube' => Schooch\Youtube\Facades\Youtube::class,
+		'FFMpeg' => Pbmedia\LaravelFFMpeg\FFMpegFacade::class,
+		'Image' => Intervention\Image\Facades\Image::class,
+		'Twitter' => Thujohn\Twitter\Facades\Twitter::class,
+		'PDF' => Barryvdh\DomPDF\Facade::class,
+		'Rekognition' => 'MichaelJWright\Rekognition\RekognitionFacade',
+		'Zipper' => Chumper\Zipper\Zipper::class,
+		'Goutte' => Weidner\Goutte\GoutteFacade::class,
 		'RedditAPI' => CodeWizz\RedditAPI\RedditAPIFacade::class,
 		'Debugbar' => Barryvdh\Debugbar\Facade::class
-    ],
+	],
 
 ];
