@@ -13,11 +13,11 @@
         </v-layout>
 
         <v-layout row wrap class="hidden-sm-and-down">
-            <v-flex xs12 sm3 md3>
+            <v-flex xs12 sm3 md4>
                 <strong>Thumbnail</strong>
             </v-flex>
 
-            <v-flex xs12 sm3 md3>
+            <v-flex xs12 sm3 md4>
                 <strong>Details</strong>
             </v-flex>
 
