@@ -1,5 +1,5 @@
 <template>
-    <v-flex xs6 sm6 md12 lg12 xl12>
+    <v-flex xs6 sm6 md12>
         <div class="thumbnail"
              :style="{backgroundImage:'url('+thumbnailImg+')'}"
              @click="onOpenDialog(asset.id)">
@@ -8,58 +8,6 @@
                 <v-icon dark medium>play_circle_outline</v-icon>
             </div>
         </div>
-
-        <!-- Image or Video in dialog -->
-        <v-dialog
-                v-model="story_dialog"
-                class="story-dialog-container"
-                content-class="story-dialog-container"
-                max-width="1200px"
-        >
-            <div class="dialog-box-switch prev">
-                <v-btn color="dark ma-0"
-                       fab
-                       small
-                       dark
-                       @click="onPreviousVideo()"
-                       :disabled="!previousImgExists">
-                    <v-icon>chevron_left</v-icon>
-                </v-btn>
-            </div>
-
-            <div class="dialog-box-switch next">
-                <v-btn color="dark ma-0"
-                       fab
-                       small
-                       dark
-                       @click="onNextVideo()"
-                       :disabled="!nextImgExists">
-                    <v-icon>chevron_right</v-icon>
-                </v-btn>
-            </div>
-
-            <v-card flat v-if="story_dialog">
-
-                <v-card-media
-                        v-if="!showVideo"
-                        :src="current_item.mime_type === 'video/mp4'? current_item.thumbnail : current_item.url"
-                        contain>
-                    <div class="video-button" v-if="current_item.mime_type === 'video/mp4'" @click="onPlayVideo()">
-                        <v-btn dark fab class="dark" medium>
-                            <v-icon dark large>play_arrow</v-icon>
-                        </v-btn>
-                    </div>
-                </v-card-media>
-
-                <v-card-media v-else>
-                    <video width="100%" height="100%" controls ref="playerVideo">
-                        <source :src="current_item.url" type="video/mp4">
-                        Your browser does not support the video tag.
-                    </video>
-                </v-card-media>
-            </v-card>
-        </v-dialog>
-
     </v-flex>
 </template>
 
@@ -67,47 +15,18 @@
     export default {
         data() {
             return {
-                loading: false,
-                loader: null,
-                current_item: '',
-                current_item_thumbnail: '',
                 thumbnailImg: '',
-
-                story_dialog: false,
-
-                nextImgExists: true,
-                nextImgObj: '',
-
-                previousImgExists: true,
-                previousImgObj: '',
-
-                showVideo: false,
             }
         },
 
-        props: [
-            'asset',
-            'story_id',
-            'assets'
-        ],
-
-        watch: {
-            loader() {
-                const l = this.loader
-                this[l] = !this[l]
-
-                setTimeout(() => (this[l] = false), 3000)
-
-                this.loader = null
-            },
-
-            story_dialog(val) {
-                if (!val) {
-                    this.showVideo = false;
-                    this.resetNextPrevious();
-                }
+        props: {
+            asset: {
+                type: Object,
+                required: true
             }
         },
+
+        watch: {},
 
         created() {
             this.setImageUrl(this.asset);
@@ -122,115 +41,8 @@
                 }
             },
 
-            downloadAsset() {
-                this.loader = 'loading';
-                this.loader = 'loading';
-                var url = '/client/stories/' + this.story_id + '/download';
-                window.location = url;
-            },
-
             onOpenDialog(id) {
-
-                this.assets.forEach((item, index) => {
-                    if (item.id === id) {
-                        this.current_item = item;
-                        this.nextImgObj = this.assets[index + 1];
-                        this.previousImgObj = this.assets[index - 1];
-
-                        if (!this.nextImgObj) {
-                            this.nextImgExists = false;
-                        }
-
-                        if (!this.previousImgObj) {
-                            this.previousImgExists = false;
-                        }
-                    }
-                })
-
-                this.story_dialog = true;
-            },
-
-
-            onPreviousVideo() {
-                this.current_item = this.previousImgObj;
-                this.showVideo = false;
-
-                this.assets.forEach((item, index) => {
-                    if (item.id == this.current_item.id) {
-                        this.current_item = item;
-                        this.nextImgObj = this.assets[index + 1];
-                        this.previousImgObj = this.assets[index - 1];
-
-                        if (!this.nextImgObj) {
-                            this.nextImgExists = false;
-                            this.previousImgExists = true;
-                        }
-
-                        else if (!this.previousImgObj) {
-                            this.previousImgExists = false;
-                            this.nextImgExists = true;
-                        }
-
-                        else {
-                            this.previousImgExists = true;
-                            this.nextImgExists = true;
-                        }
-                    }
-                })
-            },
-
-            onNextVideo() {
-                this.current_item = this.nextImgObj;
-                this.showVideo = false;
-
-                this.assets.forEach((item, index) => {
-                    if (item.id == this.current_item.id) {
-                        this.current_item = item;
-                        this.nextImgObj = this.assets[index + 1];
-                        this.previousImgObj = this.assets[index - 1];
-
-                        if (!this.nextImgObj) {
-                            this.nextImgExists = false;
-                            this.previousImgExists = true;
-                        }
-
-                        else if (!this.previousImgObj) {
-                            this.previousImgExists = false;
-                            this.nextImgExists = true;
-                        }
-
-                        else {
-                            this.previousImgExists = true;
-                            this.nextImgExists = true;
-                        }
-                    }
-                })
-            },
-
-            onCloseDialogBox() {
-                this.story_dialog = false;
-            },
-
-            onPlayVideo() {
-                var promise = new Promise((resolve, reject) => {
-                    this.showVideo = true;
-                    resolve();
-                });
-
-                promise.then(() => {
-                    setTimeout(() => {
-                        this.$refs.playerVideo.play();
-                    }, 100);
-
-                });
-
-            },
-
-            resetNextPrevious() {
-                this.nextImgExists = true;
-                this.nextImgObj = '';
-                this.previousImgExists = true;
-                this.previousImgObj = ''
+                this.$store.commit('setStoryAssetDialogBox', {open: true, id: id});
             }
         }
     }
