@@ -44,8 +44,8 @@
                                 <li>
                                     <a href="https://www.instagram.com/unilad/" target="_blank" class="instagram social-link">
                                         <svg style="width:28px;height:28px" viewBox="0 0 24 24">
-    <path fill="#ffffff" d="M7.8,2H16.2C19.4,2 22,4.6 22,7.8V16.2A5.8,5.8 0 0,1 16.2,22H7.8C4.6,22 2,19.4 2,16.2V7.8A5.8,5.8 0 0,1 7.8,2M7.6,4A3.6,3.6 0 0,0 4,7.6V16.4C4,18.39 5.61,20 7.6,20H16.4A3.6,3.6 0 0,0 20,16.4V7.6C20,5.61 18.39,4 16.4,4H7.6M17.25,5.5A1.25,1.25 0 0,1 18.5,6.75A1.25,1.25 0 0,1 17.25,8A1.25,1.25 0 0,1 16,6.75A1.25,1.25 0 0,1 17.25,5.5M12,7A5,5 0 0,1 17,12A5,5 0 0,1 12,17A5,5 0 0,1 7,12A5,5 0 0,1 12,7M12,9A3,3 0 0,0 9,12A3,3 0 0,0 12,15A3,3 0 0,0 15,12A3,3 0 0,0 12,9Z" />
-</svg>
+                                            <path fill="#ffffff" d="M7.8,2H16.2C19.4,2 22,4.6 22,7.8V16.2A5.8,5.8 0 0,1 16.2,22H7.8C4.6,22 2,19.4 2,16.2V7.8A5.8,5.8 0 0,1 7.8,2M7.6,4A3.6,3.6 0 0,0 4,7.6V16.4C4,18.39 5.61,20 7.6,20H16.4A3.6,3.6 0 0,0 20,16.4V7.6C20,5.61 18.39,4 16.4,4H7.6M17.25,5.5A1.25,1.25 0 0,1 18.5,6.75A1.25,1.25 0 0,1 17.25,8A1.25,1.25 0 0,1 16,6.75A1.25,1.25 0 0,1 17.25,5.5M12,7A5,5 0 0,1 17,12A5,5 0 0,1 12,17A5,5 0 0,1 7,12A5,5 0 0,1 12,7M12,9A3,3 0 0,0 9,12A3,3 0 0,0 12,15A3,3 0 0,0 15,12A3,3 0 0,0 12,9Z" />
+                                        </svg>
                                     </a>
                                 </li>
                             </ul>
@@ -55,6 +55,7 @@
             </v-layout>
         </v-container>
 
+        <!-- This element css display property none. so only put if you have any dialog box -->
         <div class="all-dialog-box">
             <!-- Login component -->
             <login-component></login-component>
@@ -72,15 +73,19 @@
             <!-- Story dialog box -->
             <story-dialog-component></story-dialog-component>
 
+            <!-- Story asset dialog box -->
+            <story-asset-dialog-component></story-asset-dialog-component>
+
             <!-- Buy dialog box -->
             <buy-dialog-component></buy-dialog-component>
 
             <!-- Thank you dialog box -->
             <thank-you-dialog-box></thank-you-dialog-box>
 
-            <!-- Snack bar component -->
-            <snackbar-component></snackbar-component>
         </div>
+
+        <!-- Toast bar component -->
+        <toast-component></toast-component>
 
     </section>
 </template>
@@ -90,7 +95,8 @@
     import QuoteComponent from '../component/QuoteComponent'
     import VideoDialogComponent from '../includes/VideoDialogComponent'
     import StoryDialogComponent from '../pages/clients/partials/StoryDialog'
-    import SnackbarComponent from '../includes/SnackbarComponent'
+    import StoryAssetDialogComponent from '../../../scripts/includes/StoryAssetsDialogComponent'
+    import ToastComponent from '../includes/ToastComponent'
     import BuyDialogComponent from '../component/BuyComponent'
     import ThankYouDialogBox from '../includes/ThankyouComponent'
 
@@ -114,9 +120,10 @@
             QuoteComponent,
             VideoDialogComponent,
             StoryDialogComponent,
-            SnackbarComponent,
+            ToastComponent,
             BuyDialogComponent,
-            ThankYouDialogBox
+            ThankYouDialogBox,
+            StoryAssetDialogComponent
         }
     }
 </script>
