@@ -1,8 +1,8 @@
 <template>
     <div class="client-user-create">
-        <v-container grid-list-lg class="pt-0">
+        <v-container grid-list-lg class="pt-0" v-if="iniState">
             <v-layout row wrap>
-                <v-flex xs12 pt-0 v-if="user.role !== 'client'">
+                <v-flex xs12 pt-0 v-if="$store.getters.getIsCompanyOwner">
                     <v-btn outline @click="onGoback()" class="ml-0">
                         <v-icon>chevron_left</v-icon>
                         Go back
@@ -17,90 +17,81 @@
                     <v-form ref="form" v-model="valid" id="user-create-form">
 
                         <!--User Details -->
-                        <v-container grid-list-lg>
+                        <v-layout row wrap>
+                            <v-flex xs12>
+                                <v-text-field
+                                        label="Full Name:"
+                                        v-model="user.full_name"
+                                        name="full_name"
+                                        color="dark"
+                                        :rules="[v => !!v || 'Field is required']"
+                                        required
+                                ></v-text-field>
+                            </v-flex>
+                        </v-layout>
 
-                            <v-layout row wrap>
-                                <v-flex xs12>
-                                    <v-text-field
-                                            label="Full Name:"
-                                            v-model="user.full_name"
-                                            name="full_name"
-                                            color="dark"
-                                            :rules="[v => !!v || 'Field is required']"
-                                            required
-                                    ></v-text-field>
-                                </v-flex>
-                            </v-layout>
+                        <v-layout row wrap>
+                            <v-flex xs12>
+                                <v-text-field
+                                        label="Job Title"
+                                        v-model="user.job_title"
+                                        name="job_title"
+                                        type="text"
+                                        color="dark"
+                                        :rules="[v => !!v || 'Field is required']"
+                                        required>
+                                </v-text-field>
+                            </v-flex>
 
-                            <v-layout row wrap>
-                                <v-flex xs12>
-                                    <v-text-field
-                                            label="Job Title"
-                                            v-model="user.job_title"
-                                            name="job_title"
-                                            type="text"
-                                            color="dark"
-                                            :rules="[v => !!v || 'Field is required']"
-                                            required>
-                                    </v-text-field>
-                                </v-flex>
+                            <v-flex xs12>
+                                <v-text-field
+                                        label="Email Address"
+                                        v-model="user.email"
+                                        name="email"
+                                        type="text"
+                                        color="dark"
+                                        :rules="[v => !!v || 'Field is required']"
+                                        required>
+                                </v-text-field>
+                                <small class="red--text" v-if="error && errors.email">{{ errors.email[0] }}</small>
+                            </v-flex>
 
-                                <v-flex xs12>
-                                    <v-text-field
-                                            label="Email Address"
-                                            v-model="user.email"
-                                            name="email"
-                                            type="text"
-                                            color="dark"
-                                            :rules="[v => !!v || 'Field is required']"
-                                            required>
-                                    </v-text-field>
-                                    <small class="red--text" v-if="error && errors.email">{{ errors.email[0] }}</small>
-                                </v-flex>
+                            <v-flex xs12>
+                                <v-text-field
+                                        label="Phone Number"
+                                        v-model="user.tel"
+                                        name="tel"
+                                        type="text"
+                                        color="dark"
+                                        :rules="[v => !!v || 'Field is required']"
+                                        required>
+                                </v-text-field>
+                            </v-flex>
 
-                                <v-flex xs12>
-                                    <v-text-field
-                                            label="Phone Number"
-                                            v-model="user.tel"
-                                            name="tel"
-                                            type="text"
-                                            color="dark"
-                                            :rules="[v => !!v || 'Field is required']"
-                                            required>
-                                    </v-text-field>
-                                </v-flex>
-
-                                <v-flex xs12 v-if="user.role !== 'client'">
-                                    <v-select
-                                            :items="clientRoles"
-                                            v-model="user.role"
-                                            label="Client role"
-                                            name="role"
-                                            :rules="[v => !!v || 'Field is required']"
-                                            color="dark"
-                                            item-text="name"
-                                            item-value="id"
-                                            return object
-                                            required
-                                    ></v-select>
-                                </v-flex>
-                            </v-layout>
-
-                        </v-container>
+                            <v-flex xs12 v-if="user.role !== 'client'">
+                                <v-select
+                                        :items="clientRoles"
+                                        v-model="user.role"
+                                        label="Client role"
+                                        name="role"
+                                        :rules="[v => !!v || 'Field is required']"
+                                        color="dark"
+                                        item-text="name"
+                                        item-value="id"
+                                        return object
+                                        required
+                                ></v-select>
+                            </v-flex>
+                        </v-layout>
 
                         <!-- CTA -->
-                        <v-container grid-list-lg>
-
-                            <v-layout row wrap>
-                                <v-flex xsl2 text-xs-right pa-0>
-                                    <v-btn dark
-                                           @click="onSubmit()">Update User
-                                    </v-btn>
-                                </v-flex>
-                            </v-layout>
-
-                        </v-container>
-
+                        <v-layout row wrap>
+                            <v-flex xsl2 text-xs-right pa-0>
+                                <v-btn dark
+                                       @click="onSubmit()">Update User
+                                </v-btn>
+                            </v-flex>
+                        </v-layout>
                     </v-form>
                 </v-flex>
             </v-layout>
@@ -109,19 +100,14 @@
 </template>
 
 <script>
+    import {mapGetters} from 'vuex'
+
     export default {
         data() {
             return {
                 error: false,
                 errors: null,
                 valid: false,
-                user: {
-                    full_name: null,
-                    job_title: null,
-                    email: null,
-                    tel: null,
-                    role: null,
-                },
                 emailRules: [
                     v => !!v || 'Email is required',
                     v => /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/.test(v) || 'E-mail must be valid'
@@ -130,24 +116,30 @@
                 clientRoles: [
                     {id: 'client', name: 'Client'},
                     {id: 'client_admin', name: 'Client Admin'},
-                ],
+                ]
             }
+        },
+
+        computed: {
+            ...mapGetters({
+                user: 'getCompanyCurrentUser',
+                iniState: 'getClientIniState'
+            })
         },
 
         created() {
             let slug = this.$route.params.slug;
             let userId = this.$route.params.userid;
-            axios.get('/client/profile/' + slug + '/users/' + userId + '/edit').then((response) => {
-                this.user = response.data.user;
-            });
+            let url = '/client/profile/' + slug + '/users/' + userId + '/edit';
+            this.$store.dispatch('fetchClientUser', {url: url})
         },
 
         methods: {
             onGoback() {
                 let prevRoute = this.$store.getters.getRouteUrl;
-                if(prevRoute != ''){
-                    this.$router.push({name : this.$store.getters.getRouteUrl});
-                }else{
+                if (prevRoute != '') {
+                    this.$router.push({name: this.$store.getters.getRouteUrl});
+                } else {
                     this.$router.go(-1);
                 }
             },
@@ -168,9 +160,14 @@
 
                     axios.post('/client/profile/' + slug + '/users/' + this.user.id, updateUserForm)
                         .then(response => {
-
                             if (response.data.success) {
-                                this.$router.push({name: 'client_profile'});
+                                let toast = {
+                                    message: 'User successfully updated',
+                                    duration: 3000,
+                                    color: 'success',
+                                    horizontalAlign: "right"
+                                }
+                                this.$store.commit('setToast', toast)
                             }
                         })
                         .catch(error => {
