@@ -1,9 +1,15 @@
 <template>
-    <div class="client-video-detail-section">
-        <v-container grid-list-lg class="client-story-detail-section" pt-0>
-            <v-layout row wrap v-if="story">
-                <v-flex xs12 pt-0>
-                    <v-btn outline @click="onGoback()" class="ml-0">
+    <div class="client-story-detail-section">
+        <v-container grid-list-lg
+                     pt-0
+                     class="client-story-detail-section">
+            <v-layout row wrap
+                      v-if="story">
+                <v-flex xs12
+                        pt-0>
+                    <v-btn outline
+                           class="ml-0"
+                           @click="onGoback()">
                         <v-icon>chevron_left</v-icon>
                         Go back
                     </v-btn>
@@ -11,38 +17,43 @@
 
                 <v-flex xs12 sm12 md7 lg8 xl8>
                     <div class="story-content">
-                        <v-badge right color="black" v-if="order">
-                            <span slot="badge"><v-icon dark color="white">done</v-icon></span>
+                        <v-badge right
+                                 color="black"
+                                 v-if="order">
+                            <span slot="badge">
+                                <v-icon dark color="white">done</v-icon>
+                            </span>
+
                             <h2 v-html="story.title"></h2>
                         </v-badge>
                         <h2 v-html="story.title" v-else></h2>
 
                         <div class="caption">
                             <span>Author: {{ story.author }} | </span>
-                            <span>Created at: {{ dateFormater(story.date_ingested) }}</span><br/>
-                            <!--<span>State: <strong>{{ story.state }}</strong> |</span>-->
-                            <!--<span>Status : {{ story.status }}</span>-->
+                            <span>Created at: {{ story.date_ingested | convertDate }}</span><br/>
                         </div>
 
                         <v-divider style="margin: 15px 0"></v-divider>
 
                         <div v-html="story.description"></div>
 
-                        <quote-button-component :type="'story'" :asset="story"></quote-button-component>
+                        <quote-button-component :type="'story'"
+                                                :asset="story">
+                        </quote-button-component>
                     </div>
                 </v-flex>
 
-                <v-flex xs12 sm12 md5 lg4 xl4 class="client-assets">
+                <v-flex xs12 sm12 md5 lg4
+                        class="client-assets">
                     <h2>Assets</h2>
 
-                    <v-divider style="margin-bottom:20px;"></v-divider>
+                    <v-divider class="header-divider"></v-divider>
 
                     <v-layout row wrap>
                         <asset-component v-for="asset in story.assets"
                                          :key="asset.alpha_id"
-                                         :asset="asset"
-                                         :assets="story.assets"
-                                         :story_id="story.alpha_id"></asset-component>
+                                         :asset="asset">
+                        </asset-component>
                     </v-layout>
                 </v-flex>
 
@@ -53,7 +64,6 @@
 
 <script>
     import AssetComponent from '../../../includes/StoryAssetsComponent';
-    import VideoReloadServices from '../../../services/VideoReloadServices';
     import QuoteButtonComponent from "../../../includes/BuyQuoteButtonComponent";
 
     import {mapGetters} from 'vuex';
@@ -64,7 +74,7 @@
             assetComponent: AssetComponent
         },
 
-        computed :{
+        computed: {
             ...mapGetters({
                 story: 'getCurrentStory',
                 assets: 'getCurrentStoryAssets',
@@ -82,12 +92,10 @@
 
         created() {
             this.getStoryDetail();
-
-            var video_reload = new VideoReloadServices();
         },
 
         watch: {
-            loader () {
+            loader() {
                 const l = this.loader
                 this[l] = !this[l]
 
@@ -100,21 +108,16 @@
         methods: {
             onGoback() {
                 let prevRoute = this.$store.getters.getRouteUrl;
-                if(prevRoute != ''){
-                    this.$router.push({name : this.$store.getters.getRouteUrl});
-                }else{
+                if (prevRoute != '') {
+                    this.$router.push({name: this.$store.getters.getRouteUrl});
+                } else {
                     this.$router.go(-1);
                 }
             },
 
-            getStoryDetail(){
+            getStoryDetail() {
                 let alpha_id = this.$route.params.alpha_id;
                 this.$store.dispatch('fetchCurrentStory', alpha_id);
-            },
-
-            dateFormater(date){
-                var current_date = new Date(Date.parse(date.replace('-', '/', 'g')));
-                return current_date.toDateString();
             }
         }
     }
