@@ -28,9 +28,7 @@
                 </v-layout>
             </v-container>
 
-            <search-component
-                    @searchOption="searchOption($event)"
-                    v-if="stories.length > 0"></search-component>
+            <search-component></search-component>
 
             <v-container grid-list-lg class="stories pt-0" v-if="stories.length > 0">
                 <v-layout row wrap>
