@@ -2,6 +2,13 @@
 
 \TalvBansal\MediaManager\Routes\MediaRoutes::get();
 
+//Route::get('dm', function(){
+//	//$dmResponse = Twitter::postDm(array('screen_name' => 'ianlainchbury', 'text' => 'DM Test', 'format' => 'json'));
+//	$dmResponse = Twitter::postDm(array('screen_name' => 'ianlainchbury', 'text' => 'DM Test', 'format' => 'json'));
+//
+//	dd($dmResponse);
+//});
+
 Route::group(['before' => 'if_logged_in_must_be_subscribed'], function () {
 
     Route::get('/settings_object', 'SettingController@index')->name('setting_object');
@@ -263,9 +270,9 @@ Route::group(['middleware' => ['client'], 'prefix' => 'client'], function () {
     */
 
     Route::post('collections/get_video_price/{collection_video_id}', 'CollectionController@getVideoPrice')->name('client.get_video_price');
-    Route::get('collections/accept_asset_price/{collection_asset_id}/{type}', 'CollectionController@acceptAssetQuote')->name('client.accept_asset_quote');
-    Route::get('collections/reject_asset_price/{collection_asset_id}/{type}', 'CollectionController@rejectAssetQuote')->name('client.accept_asset_quote');
-    Route::get('collections/accept_collection_quote/{collection_id}/{quote_id}', 'CollectionController@acceptCollectionQuote')->name('client.accept_collection_quote');
+    Route::post('collections/accept_asset_price/{collection_asset_id}/{type}', 'CollectionController@acceptAssetQuote')->name('client.accept_asset_quote');
+    Route::post('collections/reject_asset_price/{collection_asset_id}/{type}', 'CollectionController@rejectAssetQuote')->name('client.accept_asset_quote');
+    Route::post('collections/accept_collection_quote/{collection_id}/{quote_id}', 'CollectionController@acceptCollectionQuote')->name('client.accept_collection_quote');
     Route::post('collections/request_quote/{type}/{collection_video_id}', 'CollectionController@requestQuote')->name('client.request_quote');
 
 });
@@ -278,6 +285,7 @@ Route::group(['middleware' => ['client'], 'prefix' => 'client'], function () {
 */
 Route::post('client/collections/register_user/{collection_id}', 'CollectionController@registerUser')->name('client.register_user');
 Route::post('client/collections', 'CollectionController@store')->name('client.store');
+Route::post('client/collections/cancel_collection', 'CollectionController@cancelCollection')->name('client.cancel_collection');
 
 
 /*
